Create games.py
Number guessing game that allows users to select a range of numbers and the computer takes a random number from the batch to allow users to play.

On guessing the right number, the computer prints out a congrats message

On a wrong number,  user is given a range (whether it's high or low) to aid in finding the right answer.

Hard level where no help is given will soon be implemented.